+++ subsys/bluetooth/host/CMakeLists.txt
@@ -31,6 +31,7 @@ if(CONFIG_BT_HCI_HOST)
   zephyr_library_sources(
     uuid.c
     addr.c
+    buf.c
     hci_core.c
     hci_common.c
     )
